Perform de novo motif discovery

de_novo_motif() is a wrapper of several de novo motif discovery methods, including Streme, meme and rGADEM.

de_novo_motif(peak_input, control = "weak_peaks", method, genome_build, ...)

Arguments

peak_input

Either a path to the narrowPeak file or a GRanges peak object generated by read_peak_file().

control

The default is to use the top 0.1 peaks as input (ranked by q-value) and the remaining 0.9 peaks as the background. If you wish to supply your own control sequences, these must be passed as a DNAStringSet object.

method

De novo motif discovery method to use. The choices are "Streme", "Meme" and "rGADEM". Please note that Streme requires the MEME suite to be installed on your machine. You can find installation instructions here.

genome_build

The genome build to use. The peak sequences will be extracted from this genome.

...

Additional parameters to be passed to Streme, Meme or rGADEM. See the respective documentation for more details.

Value

The output of either Streme, Meme or rGADEM.

Details

Streme is preferable over Meme, except when you have a small number of input sequences (< 50).

Examples

if (FALSE) {
peak_file <- system.file("extdata",
                         "rep1_peaks.narrowPeak",
                         package = "MotifStats"
                         )

data("creb_motif", package = "MotifStats")

de_novo_motif(
  peak_input = peak_file,
  control = "weak_peaks",
  method = "Streme",
  genome_build = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
  )
}